feat: 커뮤니티 전체 슬라이드 적용
