+++ src/screens/(Common)/BottomTabs/Community/Community.styles.js
@@ -7,6 +7,16 @@ const styles = StyleSheet.create({
     flex: 1,
     backgroundColor: COLORS.grayscale_0,
   },
+  headerContainer: {
+    paddingTop: 24,
+    backgroundColor: COLORS.grayscale_0,
+  },
+  pager: {
+    flex: 1,
+  },
+  pagerContent: {
+    flexGrow: 1,
+  },
   listContent: {
     paddingTop: 24,
     paddingBottom: 96,
